JAVA-2657: Ability to provide a class loader to load configuration resources
This commit introduces the possibility for the user to supply a custom
class loader to locate classpath configuration resources such as
application.conf. This is particularly useful for OSGi deployments since
configuration resources typically reside inside the application bundle,
and the only loader capable of locating them is the application bundle
class loader.